DEVPROD-14637 Remove dead code from agent package
